+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Try to re-use the current syncid when showing a new screen
 * Fix `TexturedScreenHandler#transferSlot()` causing infinite loop
 * Fix `Widget`s not always getting the parent `ScreenBuilder` instance set
+* Allow `Widget`s to register themselves